Load setting from appsetting.json first version. Refactoring.
