+++ source/Roy.Logging/Roy.Logging.sln
@@ -7,7 +7,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Roy.Logging", "Roy.Logging\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Roy.UT", "Roy.UT\Roy.UT.csproj", "{5BDD0129-B90C-4EC1-B0CA-3F85CAF3C8D5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Roy.Mvc.UT", "Roy.Mvc.UT\Roy.Mvc.UT.csproj", "{704BD140-233C-4F56-A952-7AD6AE1E1C9C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Roy.Mvc.UT", "Roy.Mvc.UT\Roy.Mvc.UT.csproj", "{704BD140-233C-4F56-A952-7AD6AE1E1C9C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Roy.Logging.MVC", "Roy.Logging.MVC\Roy.Logging.MVC.csproj", "{30D5C921-74B0-44B5-919F-C0EA9D603034}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{704BD140-233C-4F56-A952-7AD6AE1E1C9C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{704BD140-233C-4F56-A952-7AD6AE1E1C9C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{704BD140-233C-4F56-A952-7AD6AE1E1C9C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{30D5C921-74B0-44B5-919F-C0EA9D603034}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{30D5C921-74B0-44B5-919F-C0EA9D603034}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{30D5C921-74B0-44B5-919F-C0EA9D603034}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{30D5C921-74B0-44B5-919F-C0EA9D603034}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
