+++ source/Roy.Logging/Roy.Mvc.UT/Program.cs
@@ -1,4 +1,7 @@
+using Microsoft.Extensions.Configuration;
+using Roy.Logging.Domain.Settings;
 using Roy.Logging.Extensions;
+using Roy.Logging.MVC.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -7,7 +10,7 @@ builder.Services.AddRazorPages();
 
 var app = builder.Build();
 
-app.UseRoyExceptionHandler();
+app.UseRoyExceptionHandler(builder, true);
 
 // Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
