+++ source/Roy.Logging/Roy.Mvc.UT/Roy.Mvc.UT.csproj
@@ -1,12 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Roy.Logging.MVC\Roy.Logging.MVC.csproj" />
     <ProjectReference Include="..\Roy.Logging\Roy.Logging.csproj" />
   </ItemGroup>
